AutoMPO

Description

ITensors.AutoMPO — Type

An AutoMPO is a data structure representing a sum of operator terms. These are products of local operators specified by names such as "Sz" or "N", times an optional coefficient which can be real or complex.

Which local operator names are available is determined by the function op associated with the TagType defined by special Index tags, such as "S=1/2","S=1", "Fermion", and "Electron".

Methods

ITensors.add! — Function

add!(ampo::AutoMPO,
     op::String, i::Int)

add!(ampo::AutoMPO,
     coef::Number,
     op1::String, i::Int)

Add a single-site operator term to the AutoMPO ampo. The operator's name is op and site number is i. The second version accepts a real or complex coefficient.

add!(ampo::AutoMPO,
     op1::String, i1::Int,
     op2::String, i2::Int,
     ops...)

add!(ampo::AutoMPO,
     coef::Number,
     op1::String, i1::Int,
     op2::String, i2::Int)

+(ampo:AutoMPO, term::Tuple)

Add a multi-site operator term to the AutoMPO ampo. Each operator is specified by a name (String) and a site number (Int). The second version accepts a real or complex coefficient.

The version involving the + operator accepts a tuple with entries either (String,Int,String,Int,...) or (Number,String,Int,String,Int,...) where these tuple values are the same as valid inputs to the add! function. For inputting a very large number of terms (tuples) to an AutoMPO, consider using the broadcasted operator .+= which avoids reallocating the AutoMPO after each addition.

Examples

julia> ampo = AutoMPO()

julia> add!(ampo,"Sz",2,"Sz",3)

julia> ampo += ("Sz",3,"Sz",4)

julia> ampo += (0.5,"S+",4,"S-",5)

julia> ampo .+= (0.5,"S+",5,"S-",6)

ITensors.MPO — Type

MPO

A finite size matrix product operator type. Keeps track of the orthogonality center.