ITensor

ITensors.ITensor — Type

An ITensor is a tensor whose interface is independent of its memory layout. Therefore it is not necessary to know the ordering of an ITensor's indices, only which indices an ITensor has. Operations like contraction and addition of ITensors automatically handle any memory permutations.

Constructors

ITensors.ITensor — Method

ITensor(iset::IndexSet)

Construct an ITensor having indices given by the IndexSet iset

ITensors.ITensor — Method

ITensor(x)

Construct a scalar ITensor with value x.

ITensor(x,i,j,...)

Construct an ITensor with indices i,j,... and all elements set to float(x).

Note that the ITensor storage will be the closest floating point version of the input value.

ITensors.randomITensor — Method

randomITensor([S,] inds)

Construct an ITensor with type S (default Float64) and indices inds, whose elements are normally distributed random numbers.

Sparse constructors

ITensors.diagITensor — Method

diagITensor(is::IndexSet)

Make a sparse ITensor of element type Float64 with non-zero elements only along the diagonal. Defaults to storing zeros along the diagonal. The storage will have Diag type.

ITensors.delta — Method

delta(::Type{T},inds::IndexSet)

Make a diagonal ITensor with all diagonal elements 1.

Operations

Base.:* — Method

*(A::ITensor, B::ITensor)

Contract ITensors A and B to obtain a new ITensor. This contraction * operator finds all matching indices common to A and B and sums over them, such that the result will have only the unique indices of A and B. To prevent indices from matching, their prime level or tags can be modified such that they no longer compare equal - for more information see the documentation on Index objects.

Base.exp — Method

exp(A::ITensor, Lis::IndexSet; hermitian = false)

Compute the exponential of the tensor A by treating it as a matrix $A_{lr}$ with the left index l running over all indices in Lis and r running over all indices not in Lis. Must have dim(Lis) == dim(inds(A))/dim(Lis) for the exponentiation to be defined. When ishermitian=true the exponential of Hermitian(A_{lr}) is computed internally.

Decompositions

LinearAlgebra.svd — Method

svd(A::ITensor,
    leftind1::Index,
    leftind2::Index,
    ...
    ;kwargs...)

Singular value decomposition (SVD) of an ITensor A, computed by treating the "left indices" provided collectively as a row index, and the remaining "right indices" as a column index (matricization of a tensor).

Whether the SVD performs a trunction depends on the keyword arguments provided. The following keyword arguments are recognized:

maxdim [Int]

mindim [Int]

cutoff [Float64]

use_absolute_cutoff [Bool] Default value: false.

use_relative_cutoff [Bool] Default value: true.

utags [String] Default value: "Link,u".

vtags [String] Default value: "Link,v".

fastsvd [Bool] Defaut value: false.