IndexSet

Missing docstring.

Missing docstring for IndexSet(::Vector{<:Index}). Check Documenter's build log for details.

Priming and tagging

Documentation for IndexSet priming and tagging methods can be found in the ITensor Priming and tagging section.

Missing docstring.

Missing docstring for map(::Function, ::IndexSet). Check Documenter's build log for details.

Set operations

Documentation for IndexSet set operations can be found in the ITensor IndexSet set operations section.

Subsets

ITensors.getfirst — Method

getfirst(f::Function, is::Indices)

Get the first Index matching the pattern function, return nothing if not found.

ITensors.getfirst — Method

getfirst(is::Indices)

Return the first Index in the Indices. If the Indices is empty, return nothing.

Missing docstring.

Missing docstring for filter(::Function, ::IndexSet). Check Documenter's build log for details.

Symmetry related properties

ITensors.dir — Method

dir(is::Indices, i::Index)

Return the direction of the Index i in the Indices is.