MPS and MPO

Types

ITensors.MPS — Type

MPS

A finite size matrix product state type. Keeps track of the orthogonality center.

ITensors.MPO — Type

MPO

A finite size matrix product operator type. Keeps track of the orthogonality center.

MPS Constructors

ITensors.MPS — Method

MPS(N::Int)

Construct an MPS with N sites with default constructed ITensors.

ITensors.MPS — Method

MPS([::Type{ElT} = Float64, ]sites)

Construct an MPS filled with Empty ITensors of type ElT from a collection of indices.

ITensors.randomMPS — Method

randomMPS(sites::Vector{<:Index}; linkdim=1)

Construct a random MPS with link dimension linkdim of type Float64.

ITensors.randomMPS — Method

randomMPS(::Type{ElT<:Number}, sites::Vector{<:Index}; linkdim=1)

Construct a random MPS with link dimension linkdim of type ElT.

ITensors.randomMPS — Method

randomMPS(sites::Vector{<:Index}, state; linkdim=1)

Construct a real, random MPS with link dimension linkdim, made by randomizing an initial product state specified by state. This version of randomMPS is necessary when creating QN-conserving random MPS (consisting of QNITensors). The initial state array provided determines the total QN of the resulting random MPS.

ITensors.productMPS — Method

productMPS(sites::Vector{<:Index},states)

Construct a product state MPS having site indices sites, and which corresponds to the initial state given by the array states. The states array may consist of either an array of integers or strings, as recognized by the state function defined for the relevant Index tag type.

Examples

N = 10
sites = siteinds("S=1/2",N)
states = [isodd(n) ? "Up" : "Dn" for n=1:N]
psi = productMPS(sites,states)

ITensors.productMPS — Method

productMPS(::Type{T},
           sites::Vector{<:Index},
           states::Union{Vector{String},
                         Vector{Int},
                         String,
                         Int})

Construct a product state MPS of element type T, having site indices sites, and which corresponds to the initial state given by the array states. The input states may be an array of strings or an array of ints recognized by the state function defined for the relevant Index tag type. In addition, a single string or int can be input to create a uniform state.

Examples

N = 10
sites = siteinds("S=1/2", N)
states = [isodd(n) ? "Up" : "Dn" for n=1:N]
psi = productMPS(ComplexF64, sites, states)
phi = productMPS(sites, "Up")

ITensors.productMPS — Method

productMPS(ivals::Vector{<:IndexVal})

Construct a product state MPS with element type Float64 and nonzero values determined from the input IndexVals.

ITensors.productMPS — Method

productMPS(::Type{T<:Number}, ivals::Vector{<:IndexVal})

Construct a product state MPS with element type T and nonzero values determined from the input IndexVals.

MPO Constructors

ITensors.MPO — Method

MPO(N::Int)

Make an MPO of length N filled with default ITensors.

ITensors.MPO — Method

MPO([::Type{ElT} = Float64}, ]sites, ops::Vector{String})

Make an MPO with pairs of sites s[i] and s[i]' and operators ops on each site.

ITensors.MPO — Method

MPO([::Type{ElT} = Float64, ]sites, op::String)

Make an MPO with pairs of sites s[i] and s[i]' and operator op on every site.

Properties

Base.length — Method

length(::MPS/MPO)

The number of sites of an MPS/MPO.

ITensors.maxlinkdim — Method

maxlinkdim(M::MPS)

maxlinkdim(M::MPO)

Get the maximum link dimension of the MPS or MPO.

ITensors.linkind — Method

linkind(M::MPS, j::Int)

linkind(M::MPO, j::Int)

Get the link or bond Index connecting the MPS or MPO tensor on site j to site j+1.

If there is no link Index, return nothing.

Priming and tagging

ITensors.prime — Method

prime(M::MPS, args...; kwargs...)

prime(M::MPO, args...; kwargs...)

Apply prime to all ITensors of an MPS/MPO, returning a new MPS/MPO.

The ITensors of the MPS/MPO will be a view of the storage of the original ITensors.

ITensors.prime! — Method

prime!(M::MPS, args...; kwargs...)

prime!(M::MPO, args...; kwargs...)

Apply prime to all ITensors of an MPS/MPO in-place.

ITensors.setprime — Method

setprime(M::MPS, args...; kwargs...)

setprime(M::MPO, args...; kwargs...)

Apply setprime to all ITensors of an MPS/MPO, returning a new MPS/MPO.

The ITensors of the MPS/MPO will be a view of the storage of the original ITensors.

ITensors.setprime! — Method

setprime!(M::MPS, args...; kwargs...)

setprime!(M::MPO, args...; kwargs...)

Apply setprime to all ITensors of an MPS/MPO in-place.

ITensors.noprime — Method

noprime(M::MPS, args...; kwargs...)

noprime(M::MPO, args...; kwargs...)

Apply noprime to all ITensors of an MPS/MPO, returning a new MPS/MPO.

The ITensors of the MPS/MPO will be a view of the storage of the original ITensors.

ITensors.noprime! — Method

noprime!(M::MPS, args...; kwargs...)

noprime!(M::MPO, args...; kwargs...)

Apply noprime to all ITensors of an MPS/MPO in-place.

ITensors.addtags — Method

addtags(M::MPS, args...; kwargs...)

addtags(M::MPO, args...; kwargs...)

Apply addtags to all ITensors of an MPS/MPO, returning a new MPS/MPO.

The ITensors of the MPS/MPO will be a view of the storage of the original ITensors.

ITensors.addtags! — Method

addtags!(M::MPS, args...; kwargs...)

addtags!(M::MPO, args...; kwargs...)

Apply addtags to all ITensors of an MPS/MPO in-place.

ITensors.removetags — Method

removetags(M::MPS, args...; kwargs...)

removetags(M::MPO, args...; kwargs...)

Apply removetags to all ITensors of an MPS/MPO, returning a new MPS/MPO.

The ITensors of the MPS/MPO will be a view of the storage of the original ITensors.

ITensors.removetags! — Method

removetags!(M::MPS, args...; kwargs...)

removetags!(M::MPO, args...; kwargs...)

Apply removetags to all ITensors of an MPS/MPO in-place.

ITensors.replacetags — Method

replacetags(M::MPS, args...; kwargs...)

replacetags(M::MPO, args...; kwargs...)

Apply replacetags to all ITensors of an MPS/MPO, returning a new MPS/MPO.

The ITensors of the MPS/MPO will be a view of the storage of the original ITensors.

ITensors.replacetags! — Method

replacetags!(M::MPS, args...; kwargs...)

replacetags!(M::MPO, args...; kwargs...)

Apply replacetags to all ITensors of an MPS/MPO in-place.

ITensors.settags — Method

settags(M::MPS, args...; kwargs...)

settags(M::MPO, args...; kwargs...)

Apply settags to all ITensors of an MPS/MPO, returning a new MPS/MPO.

The ITensors of the MPS/MPO will be a view of the storage of the original ITensors.

ITensors.settags! — Method

settags!(M::MPS, args...; kwargs...)

settags!(M::MPO, args...; kwargs...)

Apply settags to all ITensors of an MPS/MPO in-place.

Operations

ITensors.dag — Method

dag(M::MPS, args...; kwargs...)

dag(M::MPO, args...; kwargs...)

Apply dag to all ITensors of an MPS/MPO, returning a new MPS/MPO.

The ITensors of the MPS/MPO will be a view of the storage of the original ITensors.

ITensors.dag! — Method

dag!(M::MPS, args...; kwargs...)

dag!(M::MPO, args...; kwargs...)

Apply dag to all ITensors of an MPS/MPO in-place.

ITensors.orthogonalize! — Function

orthogonalize!(M::MPS, j::Int; kwargs...)
orthogonalize(M::MPS, j::Int; kwargs...)

orthogonalize!(M::MPO, j::Int; kwargs...)
orthogonalize(M::MPO, j::Int; kwargs...)

Move the orthogonality center of the MPS to site j. No observable property of the MPS will be changed, and no truncation of the bond indices is performed. Afterward, tensors 1,2,...,j-1 will be left-orthogonal and tensors j+1,j+2,...,N will be right-orthogonal.

Either modify in-place with orthogonalize! or out-of-place with orthogonalize.

NDTensors.truncate! — Function

truncate!(M::MPS; kwargs...)

truncate!(M::MPO; kwargs...)

Perform a truncation of all bonds of an MPS/MPO, using the truncation parameters (cutoff,maxdim, etc.) provided as keyword arguments.

ITensors.replacebond! — Method

replacebond!(M::MPS, b::Int, phi::ITensor; kwargs...)

Factorize the ITensor phi and replace the ITensors b and b+1 of MPS M with the factors. Choose the orthogonality with ortho="left"/"right".

ITensors.sample — Method

sample(m::MPS)

Given a normalized MPS m with orthocenter(m)==1, returns a Vector{Int} of length(m) corresponding to one sample of the probability distribution defined by squaring the components of the tensor that the MPS represents

ITensors.sample! — Method

sample!(m::MPS)

Given a normalized MPS m, returns a Vector{Int} of length(m) corresponding to one sample of the probability distribution defined by squaring the components of the tensor that the MPS represents. If the MPS does not have an orthogonality center, orthogonalize!(m,1) will be called before computing the sample.

Algebra Operations

LinearAlgebra.dot — Method

dot(A::MPS, B::MPS; make_inds_match = true)
inner(A::MPS, B::MPS; make_inds_match = true)

dot(A::MPO, B::MPO)
inner(A::MPO, B::MPO)

Compute the inner product <A|B>. If A and B are MPOs, computes the Frobenius inner product.

If make_inds_match = true, the function attempts to make the site indices match before contracting (so for example, the inputs can have different site indices, as long as they have the same dimensions or QN blocks).

For now, make_inds_match is only supported for MPSs.

See also logdot/loginner.

ITensors.logdot — Method

logdot(A::MPS, B::MPS; make_inds_match = true)
loginner(A::MPS, B::MPS; make_inds_match = true)

logdot(A::MPO, B::MPO)
loginner(A::MPO, B::MPO)

Compute the logarithm of the inner product <A|B>. If A and B are MPOs, computes the logarithm of the Frobenius inner product.

This is useful for larger MPS/MPO, where in the limit of large numbers of sites the inner product can diverge or approach zero.

If make_inds_match = true, the function attempts to make the site indices match before contracting (so for example, the inputs can have different site indices, as long as they have the same dimensions or QN blocks).

For now, make_inds_match is only supported for MPSs.

LinearAlgebra.norm — Method

norm(A::MPS)

norm(A::MPO)

Compute the norm of the MPS or MPO.

See also lognorm.

ITensors.lognorm — Method

lognorm(A::MPS)

lognorm(A::MPO)

Compute the logarithm of the norm of the MPS or MPO.

This is useful for larger MPS/MPO that are not gauged, where in the limit of large numbers of sites the norm can diverge or approach zero.

See also norm and loginner/logdot.

Base.:+ — Method

add(A::MPS, B::MPS; kwargs...)
+(A::MPS, B::MPS; kwargs...)

add(A::MPO, B::MPO; kwargs...)
+(A::MPO, B::MPO; kwargs...)

Add two MPS/MPO with each other, with some optional truncation.

Base.:* — Method

contract(::MPS, ::MPO; kwargs...)
*(::MPS, ::MPO; kwargs...)

contract(::MPO, ::MPS; kwargs...)
*(::MPO, ::MPS; kwargs...)

Contract the MPO with the MPS, returning an MPS with the unique site indices of the MPO.

Choose the method with the method keyword, for example "densitymatrix" and "naive".