IndexSet

ITensors.IndexSet — Method

IndexSet{IndexT}(inds)
IndexSet{IndexT}(inds::Index...)

Construct an IndexSet of element type IndexT from a collection of indices (any collection that is convertable to a Vector).

IndexSet(inds)
IndexSet(inds::Index...)

Construct an IndexSet from a collection of indices (any collection that is convertable to a Vector).

IndexSet(inds::Vector{<:Index})

Convert a Vector of indices to an IndexSet.

Priming and tagging

Documentation for IndexSet priming and tagging methods can be found in the ITensor Priming and tagging section.

Base.map — Method

map(f, is::IndexSet)

Apply the function to the elements of the IndexSet, returning a new IndexSet.

Set operations

Documentation for IndexSet set operations can be found in the ITensor IndexSet set operations section.

Subsets

ITensors.getfirst — Method

getfirst(f::Function, is::IndexSet)

Get the first Index matching the pattern function, return nothing if not found.

ITensors.getfirst — Method

getfirst(is::IndexSet)

Return the first Index in the IndexSet. If the IndexSet is empty, return nothing.

Base.filter — Method

filter(f::Function, inds::IndexSet)

Filter the IndexSet by the given function (output a new IndexSet with indices i for which f(i) returns true).

Note that this function is not type stable, since the number of output indices is not known at compile time.

Symmetry related properties

ITensors.dir — Method

dir(is::IndexSet, i::Index)

Return the direction of the Index i in the IndexSet is.